Make Google Cloud Build work
The issue is that the circlemator.gemspec uses `git ls-files` to
populate the list of files in the spec file. But, in GCB the working
directory is a git export and so doesn't have any git data.

This means the spec file references no files and so `rake
install:local` doesn't install anything.

Change the cloudbuild.yaml so it cleans up the working directory and
then clones the original repo so the rake install works correctly.